Validaciones

+++ microservicios/demo/src/main/java/com/example/domains/services/ActorServiceImpl.java
@@ -30,6 +30,10 @@ public class ActorServiceImpl implements ActorService {
 
 	@Override
 	public Actor add(Actor item) throws DuplicateKeyException, InvalidDataException {
+		if(item == null)
+			throw new InvalidDataException("Faltan los datos");
+		if(item.isInvalid())
+			throw new InvalidDataException("Errores de validación: ");
 		if(getOne(item.getActorId()).isPresent())
 			throw new DuplicateKeyException();
 		return dao.save(item);
@@ -37,6 +41,10 @@ public class ActorServiceImpl implements ActorService {
 
 	@Override
 	public Actor modify(Actor item) throws NotFoundException, InvalidDataException {
+		if(item == null)
+			throw new InvalidDataException("Faltan los datos");
+		if(item.isInvalid())
+			throw new InvalidDataException("Errores de validación: ");
 		if(getOne(item.getActorId()).isEmpty())
 			throw new NotFoundException();
 		return dao.save(item);
@@ -48,7 +56,9 @@ public class ActorServiceImpl implements ActorService {
 	}
 
 	@Override
-	public void delete(Actor item) {
+	public void delete(Actor item) throws InvalidDataException {
+		if(item == null)
+			throw new InvalidDataException("Faltan los datos");
 		deleteById(item.getActorId());
 	}
